requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
shapely>=2.0.0
pyclipper>=1.3.0
addict>=2.4.0
tqdm>=4.66.0
scikit-image>=0.22.0
albumentations>=1.3.1 